moved getting user results into seperate file. changed hard coded number of players to 4
