+++ client/components/App.js
@@ -10,7 +10,7 @@ const mapStateToProps = (state) => {
 }
 class App extends Component {
 
-  deal(el,props) {
+  deal(el, props) {
     let button = document.getElementById('game-button').innerHTML;
     if(props.game.dealt === false) {
       store.dispatch(types.deal())
@@ -30,12 +30,14 @@ class App extends Component {
     }
     if(button === 'Results') {
       store.dispatch(types.results());
-      // document.getElementById('game-button').innerHTML = 'Reset';
+      document.getElementById('game-button').innerHTML = 'Reset';
     }
   }
 
-  changeHands(id) {
-    store.dispatch(types.userHand(id))
+  changeHands(id, props) {
+    if(props.game.choseHandThisTurn === false) {
+      store.dispatch(types.userHand(id))
+    }
   }
   
   render() {
@@ -52,16 +54,20 @@ class App extends Component {
         </div>
         <div className="player-hands-container">
           <div>
-            <h1 onClick={(e) => this.changeHands(0)}>Hand one</h1>
+            <h1 onClick={(e) => this.changeHands(0, this.props)}>Hand one</h1>
             {this.props.game.handsDisplay[0]}
           </div>
           <div>
-            <h1 onClick={(e) => this.changeHands(1)}>Hand two</h1>
+            <h1 onClick={(e) => this.changeHands(1, this.props)}>Hand two</h1>
             {this.props.game.handsDisplay[1]}
           </div>
           <div>
-            <h1 onClick={(e) => this.changeHands(2)}>Hand three</h1>
+            <h1 onClick={(e) => this.changeHands(2, this.props)}>Hand three</h1>
             {this.props.game.handsDisplay[2]}
+          </div>
+          <div>
+            <h1 onClick={(e) => this.changeHands(3, this.props)}>Hand four</h1>
+            {this.props.game.handsDisplay[3]}
           </div>
         </div>
         <button id="game-button" onClick={(e) => this.deal(e,this.props)} >click</button>
